+++ build.zig
@@ -97,9 +97,11 @@ fn buildExecutable(b: *Build, exe: Executable, target: Build.ResolvedTarget, opt
 
     const config_mod = b.createModule(.{ .root_source_file = b.path("src/core/config.zig") });
     const pid_mod = b.createModule(.{ .root_source_file = b.path("src/core/pid.zig") });
+    const net_mod = b.createModule(.{ .root_source_file = b.path("src/core/network.zig") });
 
     mod.addImport("config", config_mod);
     mod.addImport("pid", pid_mod);
+    mod.addImport("network", net_mod);
     mod.addImport("build_options", build_options.createModule());
 
     const obj = b.addExecutable(.{