+++ src/modules/ping.zig
@@ -32,6 +32,7 @@ const Allocator = mem.Allocator;
 
 const pid = @import("pid");
 const config = @import("config");
+const network = @import("network");
 
 const PingError = error{
     Timeout,
@@ -114,33 +115,6 @@ noinline fn ping(buffer: []u8, data: Data) !i64 {
     return latency;
 }
 
-pub fn resolveIP(allocator: Allocator, domain: []const u8, port: []const u8) !?[]const u8 {
-    const domain_cstr = try allocator.dupeZ(u8, domain);
-    const port_cstr = try allocator.dupeZ(u8, port);
-
-    const hints = c.addrinfo{
-        .family = posix.AF.INET,
-        .socktype = posix.SOCK.STREAM,
-        .protocol = 0,
-        .flags = c.AI{},
-        .addrlen = 0,
-        .canonname = null,
-        .addr = null,
-        .next = null,
-    };
-
-    var result: ?*c.addrinfo = null;
-    _ = c.getaddrinfo(domain_cstr.ptr, port_cstr.ptr, &hints, &result);
-    defer c.freeaddrinfo(result.?);
-
-    const node = result.?;
-
-    const ipv4 = @as(*const posix.sockaddr.in, @ptrCast(@alignCast(node.addr.?)));
-    const bytes = mem.asBytes(&ipv4.addr);
-
-    return try fmt.allocPrint(allocator, "{}.{}.{}.{}", .{ bytes[0], bytes[1], bytes[2], bytes[3] });
-}
-
 const UpdateIPArguments = struct {
     allocator: Allocator,
     data: *Data,
@@ -148,7 +122,7 @@ const UpdateIPArguments = struct {
 
 fn updateIP(args: UpdateIPArguments) !void {
     while (true) {
-        if (try resolveIP(args.allocator, args.data.TARGET_DOMAIN, args.data.TARGET_PORT)) |ip| args.data.TARGET_IP = ip;
+        if (try network.resolveIP(args.allocator, args.data.TARGET_DOMAIN, args.data.TARGET_PORT)) |ip| args.data.TARGET_IP = ip;
         Thread.sleep(@intCast(args.data.TARGET_UPDATE_MS * time.ns_per_s));
     }
 }